requirements.txt
streamlit>=1.36
replicate>=0.25.0
gTTS>=2.5.0
moviepy>=1.0.3
pydub>=0.25.1
python-dotenv>=1.0.1
tenacity>=8.2.3
requests>=2.31.0
pyttsx3>=2.90
